ui: move rename namespace to component
This commit removes the part of the code responsible for renaming the
namespace to a new component responsible for just that functionality.
