+++ ui/src/components/setting/SettingNamespace.vue
@@ -57,65 +57,7 @@
             class="mt-6"
             data-test="editOperation-div"
           >
-            <ValidationObserver
-              ref="obs"
-              v-slot="{ passes }"
-            >
-              <v-row>
-                <v-col>
-                  <h3>
-                    Namespace
-                  </h3>
-                </v-col>
-
-                <v-spacer />
-
-                <v-col
-                  md="auto"
-                  class="ml-auto"
-                >
-                  <v-tooltip
-                    bottom
-                    :disabled="hasAuthorizationRenameNamespace"
-                  >
-                    <template #activator="{ on }">
-                      <div v-on="on">
-                        <v-btn
-                          :disabled="!hasAuthorizationRenameNamespace"
-                          color="primary"
-                          @click="passes(editNamespace)"
-                        >
-                          Rename Namespace
-                        </v-btn>
-                      </div>
-                    </template>
-
-                    <span>
-                      You don't have this kind of authorization.
-                    </span>
-                  </v-tooltip>
-                </v-col>
-              </v-row>
-
-              <div class="mt-4 mb-2">
-                <ValidationProvider
-                  v-slot="{ errors }"
-                  ref="providerName"
-                  vid="name"
-                  name="Priority"
-                  rules="required|rfc1123|noDot|namespace"
-                >
-                  <v-text-field
-                    v-model="name"
-                    class="ml-3"
-                    label="Name"
-                    :error-messages="errors"
-                    required
-                    data-test="name-text"
-                  />
-                </ValidationProvider>
-              </div>
-            </ValidationObserver>
+            <NamespaceRename data-test="namespaceRename-component" />
 
             <v-divider />
             <v-divider />
@@ -206,35 +148,25 @@
 
 <script>
 
-import {
-  ValidationObserver,
-  ValidationProvider,
-} from 'vee-validate';
-
 import SettingSecurity from '@/components/setting/SettingSecurity';
 import NamespaceMemberList from '@/components/namespace/NamespaceMemberList';
+import NamespaceRename from '@/components/namespace/NamespaceRename';
 import NamespaceMemberFormDialog from '@/components/namespace/NamespaceMemberFormDialog';
 import NamespaceDelete from '@/components/namespace/NamespaceDelete';
-
-import hasPermission from '@/components/filter/permission';
 
 export default {
   name: 'SettingNamespaceComponent',
 
-  filters: { hasPermission },
-
   components: {
-    ValidationProvider,
-    ValidationObserver,
     NamespaceDelete,
     NamespaceMemberList,
+    NamespaceRename,
     NamespaceMemberFormDialog,
     SettingSecurity,
   },
 
   data() {
     return {
-      name: '',
       namespaceMemberFormShow: false,
     };
   },
@@ -251,18 +183,6 @@ export default {
     isEnterprise() {
       return this.$env.isEnterprise;
     },
-
-    hasAuthorizationRenameNamespace() {
-      const role = this.$store.getters['auth/role'];
-      if (role !== '') {
-        return hasPermission(
-          this.$authorizer.role[role],
-          this.$actions.namespace.rename,
-        );
-      }
-
-      return false;
-    },
   },
 
   async created() {
@@ -276,30 +196,9 @@ export default {
       this.getNamespace();
     },
 
-    async editNamespace() {
-      try {
-        await this.$store.dispatch('namespaces/put', { id: this.tenant, name: this.name });
-        await this.$store.dispatch('namespaces/get', this.tenant);
-        this.$store.dispatch('snackbar/showSnackbarSuccessAction', this.$success.namespaceEdit);
-      } catch (error) {
-        if (error.response.status === 400) {
-          this.$refs.obs.setErrors({
-            namespace: this.$errors.form.invalid('namespace', 'nonStandardCharacters'),
-          });
-        } else if (error.response.status === 409) {
-          this.$refs.obs.setErrors({
-            namespace: this.$errors.form.invalid('namespace', 'nameUsed'),
-          });
-        } else {
-          this.$store.dispatch('snackbar/showSnackbarErrorAction', this.$errors.snackbar.namespaceEdit);
-        }
-      }
-    },
-
     async getNamespace() {
       try {
         await this.$store.dispatch('namespaces/get', this.tenant);
-        this.name = this.namespace.name;
       } catch (error) {
         if (error.response.status === 403) {
           this.$store.dispatch('snackbar/showSnackbarErrorAssociation');
