+++ ui/tests/unit/components/setting/SettingNamespace.spec.js
@@ -1,29 +1,17 @@
 import Vuex from 'vuex';
-import { mount, createLocalVue } from '@vue/test-utils';
-import { ValidationProvider, ValidationObserver } from 'vee-validate';
-import flushPromises from 'flush-promises';
+import { shallowMount, createLocalVue } from '@vue/test-utils';
 import Vuetify from 'vuetify';
 import SettingNamespace from '@/components/setting/SettingNamespace';
 import '@/vee-validate';
-import { actions, authorizer } from '../../../../src/authorizer';
 
 describe('SettingNamespace', () => {
   const localVue = createLocalVue();
   const vuetify = new Vuetify();
   localVue.use(Vuex);
-  localVue.component('ValidationProvider', ValidationProvider);
-  localVue.component('ValidationObserver', ValidationObserver);
 
   document.body.setAttribute('data-app', true);
 
   let wrapper;
-
-  const role = ['owner', 'operator'];
-
-  const hasAuthorizationRenameNamespace = {
-    owner: true,
-    operator: false,
-  };
 
   const members = [
     {
@@ -49,15 +37,6 @@ describe('SettingNamespace', () => {
 
   const hostedNamespace = { ...openNamespace, max_devices: -1 };
 
-  const invalidNamespaces = [
-    '\'', '"', '!', '@', '#', '$', '%', '¨', '&', '*', '(', ')', '-', '_', '=', '+', '´', '`', '[',
-    '{', '~', '^', ']', ',', '<', '..', '>', ';', ':', '/', '?',
-  ];
-
-  const invalidMinAndMaxCharacters = [
-    's', 'sh', 'xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx', 'xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx',
-  ];
-
   const tests = [
     {
       description: 'Open version',
@@ -69,7 +48,7 @@ describe('SettingNamespace', () => {
         isEnterprise: false,
       },
       data: {
-        name: openNamespace.name,
+        namespaceMemberFormShow: false,
       },
       computed: {
         namespace: openNamespace,
@@ -77,6 +56,7 @@ describe('SettingNamespace', () => {
         isEnterprise: false,
       },
       components: {
+        'namespaceRename-component': true,
         'namespaceMemberFormDialogAdd-component': true,
         'namespaceDelete-component': true,
       },
@@ -99,7 +79,7 @@ describe('SettingNamespace', () => {
         isEnterprise: true,
       },
       data: {
-        name: hostedNamespace.name,
+        namespaceMemberFormShow: false,
       },
       computed: {
         namespace: hostedNamespace,
@@ -107,6 +87,7 @@ describe('SettingNamespace', () => {
         isEnterprise: true,
       },
       components: {
+        'namespaceRename-component': true,
         'namespaceMemberFormDialogAdd-component': true,
         'namespaceDelete-component': true,
       },
@@ -121,24 +102,19 @@ describe('SettingNamespace', () => {
     },
   ];
 
-  const storeVuex = (namespace, authID, tenant, currentrole) => new Vuex.Store({
+  const storeVuex = (namespace, authID, tenant) => new Vuex.Store({
     namespaced: true,
     state: {
       namespace,
       authID,
       tenant,
-      currentrole,
     },
     getters: {
       'namespaces/get': (state) => state.namespace,
-      'auth/id': (state) => state.id,
       'auth/tenant': (state) => state.tenant,
-      'auth/role': (state) => state.currentrole,
     },
     actions: {
-      'namespaces/put': () => {},
       'namespaces/get': () => {},
-      'security/get': () => {},
       'namespaces/removeUser': () => {},
       'snackbar/showSnackbarSuccessAction': () => {},
       'snackbar/showSnackbarErrorAction': () => {},
@@ -147,150 +123,85 @@ describe('SettingNamespace', () => {
   });
 
   tests.forEach((test) => {
-    role.forEach((currentrole) => {
-      describe(`${test.description} ${currentrole}`, () => {
-        beforeEach(async () => {
-          jest.spyOn(Storage.prototype, 'getItem').mockReturnValue('e359bf484715');
+    describe(`${test.description}`, () => {
+      beforeEach(async () => {
+        jest.spyOn(Storage.prototype, 'getItem').mockReturnValue('e359bf484715');
 
-          wrapper = mount(SettingNamespace, {
-            store: storeVuex(
-              test.variables.namespace,
-              test.variables.authID,
-              test.variables.tenant,
-              currentrole,
-            ),
-            localVue,
-            stubs: ['fragment'],
-            vuetify,
-            mocks: {
-              $authorizer: authorizer,
-              $actions: actions,
-              $env: {
-                isEnterprise: test.variables.isEnterprise,
-              },
-              $stripe: {
-                elements: () => ({
-                  create: () => ({
-                    mount: () => null,
-                  }),
-                }),
-              },
+        wrapper = shallowMount(SettingNamespace, {
+          store: storeVuex(
+            test.variables.namespace,
+            test.variables.authID,
+            test.variables.tenant,
+          ),
+          localVue,
+          stubs: ['fragment'],
+          vuetify,
+          mocks: {
+            $env: {
+              isEnterprise: test.variables.isEnterprise,
             },
-          });
+            $stripe: {
+              elements: () => ({
+                create: () => ({
+                  mount: () => null,
+                }),
+              }),
+            },
+          },
         });
+      });
 
-        ///////
-        // Component Rendering
-        //////
+      ///////
+      // Component Rendering
+      //////
 
-        it('Is a Vue instance', () => {
-          expect(wrapper).toBeTruthy();
+      it('Is a Vue instance', () => {
+        expect(wrapper).toBeTruthy();
+      });
+      it('Renders the component', () => {
+        expect(wrapper.html()).toMatchSnapshot();
+      });
+
+      ///////
+      // Data checking
+      //////
+
+      it('Compare data with default value', () => {
+        Object.keys(test.data).forEach((item) => {
+          expect(wrapper.vm[item]).toEqual(test.data[item]);
         });
-        it('Renders the component', () => {
-          expect(wrapper.html()).toMatchSnapshot();
+      });
+      it('Process data in the computed', () => {
+        Object.keys(test.computed).forEach((item) => {
+          expect(wrapper.vm[item]).toEqual(test.computed[item]);
         });
+      });
+      it('Process data in methods', () => {
+        let percent = 0;
+        if (test.variables.namespace.max_devices >= 0) {
+          percent = (wrapper.vm.countDevicesHasNamespace()
+            / test.variables.namespace.max_devices) * 100;
+        }
 
-        ///////
-        // Data checking
-        //////
+        expect(wrapper.vm.hasTenant()).toEqual(test.variables.hasTenant);
+        expect(wrapper.vm.countDevicesHasNamespace())
+          .toEqual(test.variables.namespace.devices_count);
+        expect(wrapper.vm.countDevicesHasNamespacePercent())
+          .toEqual({ maxDevices: test.variables.namespace.max_devices, percent });
+      });
 
-        it('Compare data with default value', () => {
-          Object.keys(test.data).forEach((item) => {
-            expect(wrapper.vm[item]).toEqual(test.data[item]);
-          });
+      //////
+      // HTML validation
+      //////
+
+      it('Renders the template with components', () => {
+        Object.keys(test.components).forEach((item) => {
+          expect(wrapper.find(`[data-test="${item}"]`).exists()).toBe(test.components[item]);
         });
-        it('Process data in the computed', () => {
-          Object.keys(test.computed).forEach((item) => {
-            expect(wrapper.vm[item]).toEqual(test.computed[item]);
-          });
-          expect(wrapper.vm.hasAuthorizationRenameNamespace)
-            .toEqual(hasAuthorizationRenameNamespace[currentrole]);
-        });
-        it('Process data in methods', () => {
-          let percent = 0;
-          if (test.variables.namespace.max_devices >= 0) {
-            percent = (wrapper.vm.countDevicesHasNamespace()
-              / test.variables.namespace.max_devices) * 100;
-          }
-
-          expect(wrapper.vm.hasTenant()).toEqual(test.variables.hasTenant);
-          expect(wrapper.vm.countDevicesHasNamespace())
-            .toEqual(test.variables.namespace.devices_count);
-          expect(wrapper.vm.countDevicesHasNamespacePercent())
-            .toEqual({ maxDevices: test.variables.namespace.max_devices, percent });
-        });
-
-        //////
-        // HTML validation
-        //////
-
-        it('Renders the template with components', () => {
-          Object.keys(test.components).forEach((item) => {
-            expect(wrapper.find(`[data-test="${item}"]`).exists()).toBe(test.components[item]);
-          });
-        });
-        it('Renders the template with data', () => {
-          Object.keys(test.template).forEach((item) => {
-            expect(wrapper.find(`[data-test="${item}"]`).exists()).toBe(test.template[item]);
-          });
-        });
-
-        //////
-        // In this case, invalid RFC1123.
-        //////
-
-        invalidNamespaces.forEach((inamespace) => {
-          it(`Shows invalid namespace error for ${inamespace}`, async () => {
-            wrapper.setData({ name: inamespace });
-            await flushPromises();
-
-            const validator = wrapper.vm.$refs.providerName;
-
-            await validator.validate();
-            expect(validator.errors[0]).toBe('You entered an invalid RFC1123 name');
-          });
-        });
-
-        //////
-        // In this case, password should be 3-30 characters long.
-        //////
-
-        invalidMinAndMaxCharacters.forEach((character) => {
-          it(`Shows invalid namespace error for ${character}`, async () => {
-            wrapper.setData({ name: character });
-            await flushPromises();
-
-            const validator = wrapper.vm.$refs.providerName;
-
-            await validator.validate();
-            expect(validator.errors[0]).toBe('Your namespace should be 3-30 characters long');
-          });
-        });
-
-        it('Show validation messages', async () => {
-          //////
-          // In this case, validate fields required.
-          //////
-
-          wrapper.setData({ name: '' });
-          await flushPromises();
-
-          let validator = wrapper.vm.$refs.providerName;
-
-          await validator.validate();
-          expect(validator.errors[0]).toBe('This field is required');
-
-          //////
-          // In this case, must not contain dots.
-          //////
-
-          wrapper.setData({ name: 'ShelHub.' });
-          await flushPromises();
-
-          validator = wrapper.vm.$refs.providerName;
-
-          await validator.validate();
-          expect(validator.errors[0]).toBe('The name must not contain dots');
+      });
+      it('Renders the template with data', () => {
+        Object.keys(test.template).forEach((item) => {
+          expect(wrapper.find(`[data-test="${item}"]`).exists()).toBe(test.template[item]);
         });
       });
     });
